move from adapter to scorer
